# Added flag to disable l2-dist finite check