--- a/sklearn/metrics/tests/test_pairwise.py
+++ b/sklearn/metrics/tests/test_pairwise.py
@@ -718,7 +718,8 @@ def test_euclidean_no_check_inputs_with_bad_input():
     # Ensure failure when check_inputs is False and the inputs are bad
     X = [[0]]
     Y = [[1], [2]]
-    assert_raises(AttributeError, euclidean_distances, X, Y, check_inputs=False)
+    assert_raises(AttributeError, euclidean_distances, X, Y,
+                  check_inputs=False)
 
     # Unsigned integer types might work if squared is False
     # and numpy casting rule is not 'same_kind'
@@ -814,7 +815,8 @@ def test_pairwise_no_check_inputs_with_good_input():
         if metric in PAIRWISE_DISTANCE_FUNCTIONS:
             # Check the pairwise_distances implementation
             # gives the same value
-            distances = PAIRWISE_DISTANCE_FUNCTIONS[metric](X, Y, check_inputs=False)
+            distances = PAIRWISE_DISTANCE_FUNCTIONS[metric](X, Y,
+                                                            check_inputs=False)
             distances = np.diag(distances)
             assert_array_almost_equal(distances, S2)
 
@@ -828,9 +830,11 @@ def test_pairwise_no_check_inputs_with_bad_input():
     Y = rng.random_sample((6, 4)).tolist()
     for metric, func in iteritems(PAIRED_DISTANCES):
         assert_raises((ValueError, TypeError), func, X, Y, check_inputs=False)
-        assert_raises(ValueError, func, csr_matrix(X), csr_matrix(Y), check_inputs=False)
+        assert_raises(ValueError, func, csr_matrix(X), csr_matrix(Y),
+                      check_inputs=False)
         if metric in PAIRWISE_DISTANCE_FUNCTIONS:
             # Check the pairwise_distances implementation
             # many different types of errors can happen here
             assert_raises((ValueError, AttributeError),
-                          PAIRWISE_DISTANCE_FUNCTIONS[metric], X, Y, check_inputs=False)
+                          PAIRWISE_DISTANCE_FUNCTIONS[metric], X, Y,
+                          check_inputs=False)