@amueller 2016-09-09
please address the flake8 errors.
I'm not sure if we should add a flag for finiteness specifically. We could also just entirely skip ``check_array`` and add a ``check_input`` flag. That seems conceptually more simple.