@GaelVaroquaux 2016-09-10
Great work!

I am in favor of @amueller 's suggestion: skipping completely the check_array in the inner loop, rather than adding a flag to check_array.