@Erotemic 2016-09-12
Sounds good. I'll make those changes and push.