@Erotemic 2016-09-13
I finished these changes. Let me know if anything looks like it needs more work.