@amueller 2016-09-13
LGTM. Can you please re-run your benchmarks with the current version and check that the improvement over master is still there? (it's easy to mess up things when refactoring ;)