@Erotemic 2016-09-14
I ran the benchmarks on a smaller grid because the older grid took a long time. This one still shows the same speed improvements. I also added a parameter to test giving integer inputs.

```
    n_features  asint  per_cluster  n_clusters  new_speed  old_speed  percent_change  absolute_change
0          128   True           20         500   1.728701   3.175227       45.556624         1.446526
1          128   True           10         500   0.964327   1.461100       33.999946         0.496773
18         128  False           20         500   1.820726   2.117415       14.011850         0.296689
19         128  False           10         500   0.737921   1.026572       28.117949         0.288651
3           32   True           20         500   0.681702   0.790321       13.743655         0.108619
22          32  False           10         500   0.361195   0.459011       21.310159         0.097816
5           32   True            1         500   0.117711   0.205754       42.790332         0.088043
4           32   True           10         500   0.428249   0.515482       16.922577         0.087233
20         128  False            1         500   0.173600   0.255813       32.137977         0.082213
2          128   True            1         500   0.209229   0.289853       27.815446         0.080624
25         128  False           10         100   0.069253   0.129692       46.601908         0.060439
23          32  False            1         500   0.105972   0.155374       31.795527         0.049402
21          32  False           20         500   0.687639   0.727779        5.515399         0.040140
7          128   True           10         100   0.047142   0.085005       44.541926         0.037863
27          32  False           20         100   0.064948   0.098792       34.257885         0.033844
24         128  False           20         100   0.076329   0.098159       22.239494         0.021830
10          32   True           10         100   0.056360   0.077750       27.511216         0.021390
9           32   True           20         100   0.075062   0.096348       22.092830         0.021286
6          128   True           20         100   0.122499   0.142712       14.163424         0.020213
26         128  False            1         100   0.023037   0.041862       44.969302         0.018825
8          128   True            1         100   0.038797   0.053073       26.898888         0.014276
11          32   True            1         100   0.008956   0.020436       56.175699         0.011480
29          32  False            1         100   0.013403   0.021140       36.597607         0.007737
28          32  False           10         100   0.028573   0.034658       17.557063         0.006085
15          32   True           20          10   0.001318   0.002301       42.720962         0.000983
13         128   True           10          10   0.001378   0.002235       38.346667         0.000857
35          32  False            1          10   0.001117   0.001886       40.771176         0.000769
16          32   True           10          10   0.001199   0.001932       37.936567         0.000733
33          32  False           20          10   0.001339   0.002054       34.811376         0.000715
14         128   True            1          10   0.001133   0.001839       38.397718         0.000706
34          32  False           10          10   0.001206   0.001889       36.147924         0.000683
17          32   True            1          10   0.001096   0.001775       38.253862         0.000679
32         128  False            1          10   0.001167   0.001811       35.558189         0.000644
12         128   True           20          10   0.002095   0.002675       21.677511         0.000580
30         128  False           20          10   0.002064   0.002503       17.536674         0.000439
31         128  False           10          10   0.007326   0.002133     -243.471943        -0.005193
```

The modified benchmark script is in this gist https://gist.github.com/Erotemic/5230d93ccc9fa5329b0a02a351b02939