@Erotemic 2016-09-15
I'm not sure why these errors are happening on travis. The last message complains about subprocess which is odd considering it is a global import in setup.py. The tests all pass on my machine, so I'm having trouble pinpointing the problem. 

```
Partial import of sklearn during the build process.
Generating cython files
Cythonizing sources
Traceback (most recent call last):
  File "setup.py", line 304, in <module>
    setup_package()
  File "setup.py", line 287, in setup_package
    generate_cython()
  File "setup.py", line 186, in generate_cython
    p = subprocess.call([sys.executable, os.path.join(cwd,
NameError: name 'subprocess' is not defined
```