@Erotemic 2016-10-10
I made a few more changes to this branch. Here is a summary of the differences. 

**FIX: logic error in euclidean_distances**

I noticed a small issue in euclidean_distances when `check_inputs` is False. I had incorrect indentation leading to a NameError because `XX` was undefined. I fixed this and added a new test to cover this case.

**FIX: documentation error in euclidean_distances**

When fixing this I also noticed that the documentation states that the input shape of `X_norm_squared` should be `(1, n_samples_1)`, but the underlying function actually wants it to be `(n_samples_1, 1)`. If check_inputs is True the underlying code will correct itself, but with the new `check_inputs=False` behavior the documentation needs an update. I corrected this in the documentation. 


**ADD: check_flags to other pairwise metrics functions**

I also went ahead and added the `check_inputs` flag to the other metrics in pairwise.py in order to maintain a consistent API. The only function I skipped was `paired_distances` because I couldn't see how to nicely work with its callable version without using kwds in complicated ways. I added a total of 4 new tests to cover all of this new behavior. 

**FIX: Converted one test in test_pairwise.py to non-yield version**

While I was in that file I changed the behavior of a yield test which was producing a lot of warning on my system with py.test version 3.0.1. It seems like this behavior will be deprecated in a new version, so
it may be worth the effort to go through and change those tests to their non-yield versions soon. 
 
---
I'm still waiting on the results of AppVeyor, but it seems as if all tests are now passing in this branch.