@Erotemic 2016-10-12
I dropped the commit that changes the yield test behavior and made a new PR #7654 which looks at the issue independently.