@jnothman 2016-11-01
@Erotemic, will #7548 (perhaps with a context manager) fix this in your opinion?